+++ src/main/java/com/LMS/LMS/ModelLayer/Assignment.java
@@ -5,14 +5,15 @@ import java.util.List;
 
 @Entity
 public class Assignment {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String title;
     private String description;
     private LocalDateTime dueDate;
-    private Integer maxScore;
+    private Long maxScore;
 
     @ManyToOne
     private Course course;
@@ -23,8 +24,8 @@ public class Assignment {
     // Constructors
     public Assignment() {}
 
-    public Assignment(Integer id, String title, String description, LocalDateTime dueDate,
-                      Integer maxScore, Course course) {
+    public Assignment(Long id, String title, String description, LocalDateTime dueDate,
+                      Long maxScore, Course course) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -34,8 +35,8 @@ public class Assignment {
     }
 
     // Getters and Setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
@@ -46,8 +47,8 @@ public class Assignment {
     public LocalDateTime getDueDate() { return dueDate; }
     public void setDueDate(LocalDateTime dueDate) { this.dueDate = dueDate; }
 
-    public Integer getMaxScore() { return maxScore; }
-    public void setMaxScore(Integer maxScore) { this.maxScore = maxScore; }
+    public Long getMaxScore() { return maxScore; }
+    public void setMaxScore(Long maxScore) { this.maxScore = maxScore; }
 
     public Course getCourse() { return course; }
     public void setCourse(Course course) { this.course = course; }
