+++ src/main/java/com/LMS/LMS/ModelLayer/Attendance.java
@@ -6,7 +6,7 @@ public class Attendance {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     private User student;
@@ -16,7 +16,7 @@ public class Attendance {
 
     private boolean attend;
 
-    public Attendance(Integer id, User student, Lesson lesson, boolean attend) {
+    public Attendance(Long id, User student, Lesson lesson, boolean attend) {
         this.id = id;
         this.student = student;
         this.lesson = lesson;
@@ -26,11 +26,11 @@ public class Attendance {
     public Attendance() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
