+++ src/main/java/com/LMS/LMS/RepositoryLayer/CourseRepository.java
@@ -3,5 +3,5 @@ package com.LMS.LMS.RepositoryLayer;
 import com.LMS.LMS.ModelLayer.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
 }
