+++ src/main/java/com/LMS/LMS/RepositoryLayer/LessonRepository.java
@@ -3,5 +3,5 @@ package com.LMS.LMS.RepositoryLayer;
 import com.LMS.LMS.ModelLayer.Lesson;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LessonRepository extends JpaRepository<Lesson, Integer> {
+public interface LessonRepository extends JpaRepository<Lesson, Long> {
 }
